Working Python Version / Halted C++ Version
Got stuck on the object oriented aspects of C++ as I've done most of my recent coding in Python. Adapted what we made in C++ to python and have a working version using random integers to decide who gets what next instead of set start / finish dates for circulation. In doing so I'm assuming that people are free to hold onto a book for as long as they want to, even if they pass it along very timely.